requirements.txt
pandas
pyarrow
boto3
trino
python-dotenv
faker